*** empty log message ***
svn path=/trunk/kdelibs/; revision=173987

+++ kio/kio/kdirlister_p.h
@@ -218,7 +218,7 @@ private:
 
     ~DirItem()
     {
-      if ( autoUpdates )
+      if ( autoUpdates && KDirWatch::exists() )
         kdirwatch->removeDir( url.path() );
 
       delete rootItem;
